:recycle: Rename toolbox dropdownV2 to dropdown
